fix: 删除一些没用的文件
