+++ demo3/config/paths.js
@@ -44,7 +44,6 @@ module.exports = {
   appBuild: resolveApp('build'),
   appPublic: resolveApp('public'),
   appHtml: resolveApp('public/index.html'),
-  appIndexJs: resolveApp('src/index.js'),
   TtReduxJs: resolveApp('src/TtRedux.js'),
   appPackageJson: resolveApp('package.json'),
   appSrc: resolveApp('src'),
