+++ demo3/package.json
@@ -35,6 +35,8 @@
     "react": "^16.2.0",
     "react-dev-utils": "^5.0.0",
     "react-dom": "^16.2.0",
+    "react-redux": "^5.0.7",
+    "redux": "^3.7.2",
     "redux-saga": "^0.16.0",
     "style-loader": "0.19.0",
     "sw-precache-webpack-plugin": "0.11.4",
